Exposing Kubevirt dynamic plugin missing components
